requirements.txt
aniso8601==9.0.1
APScheduler==3.8.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-APScheduler==1.12.3
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
protobuf==3.19.4
pycodestyle==2.8.0
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.0
six==1.16.0
SQLAlchemy==1.4.29
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
Werkzeug==2.0.2
